requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
Django==6.0.3
django-admin-interface==0.32.0
django-colorfield==0.14.0
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.11
importlib_resources==6.5.2
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.10
requests==2.32.5
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
WMI==1.5.1
waitress

